% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content_data.R
\name{get_content_data}
\alias{get_content_data}
\title{get_content_data}
\usage{
get_content_data(
  customer_id,
  customer_list,
  page,
  type,
  start,
  end,
  total_pages = FALSE
)
}
\arguments{
\item{customer_id}{customer_id returned from get_customer_id function, used in json payload}

\item{customer_list}{list of profile_ids for all accounts, used in json payload}

\item{page}{page number for data to pull, used in json payload}

\item{type}{whether you're pulling profiles or posts data}

\item{start}{start date, used in json payload}

\item{end}{end date, used in json payload}

\item{total_pages}{option to also return the total number of pages of data for looping purposes. Defaults to FALSE}
}
\description{
Return contents of API call for one page worth of data. Uses global api_token created by sprout_auth()
}
