% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customer_id.R
\name{get_customer_id}
\alias{get_customer_id}
\title{get_customer_id}
\usage{
get_customer_id()
}
\description{
Determine a members customer_id which is needed for future API calls. Uses global api_token created by sprout_auth()
}
