% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customer_list.R
\name{get_customer_list}
\alias{get_customer_list}
\title{get_customer_list}
\usage{
get_customer_list(customer_id)
}
\arguments{
\item{customer_id}{customer_id returned from get_customer_id function}
}
\value{
a list of the profile ids
}
\description{
Pull the list of profile_ids for all accounts to use in the body of future requests
}
