% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprout_auth.R
\name{sprout_auth}
\alias{sprout_auth}
\title{sprout_auth}
\usage{
sprout_auth(token)
}
\arguments{
\item{token}{API key, can be passed via SYS.getenv}
}
\description{
Function that takes in an API key and formats it for future calls and stores it as a global variable
}
