% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprout_encourage.R
\name{sprout_encourage}
\alias{sprout_encourage}
\title{sprout_encourage}
\usage{
sprout_encourage()
}
\value{
A character string with a randomly chosen motivational message.
}
\description{
Returns a friendly message to encourage learners who are just getting started with R.
}
\examples{
sprout_encourage()
}
