% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customer_id.R
\name{get_customer_id}
\alias{get_customer_id}
\title{get_customer_id}
\usage{
get_customer_id(token)
}
\arguments{
\item{token}{API key}
}
\description{
Determine a members customer_id which is needed for future API calls
}
